chatbot-app/backend/requirement.txt
Flask
flask-cors
python-dotenv
requests
uvicorn
fastapi
